// src/data/largeObjectGraphics.gen.ts
// Auto-generated from pokeemerald object event graphics C source.
// Do not edit manually. Run: npm run generate:large-objects

import type { LargeObjectGraphicsId } from '../types/objectEvents';

export interface LargeObjectGraphicsInfo {
  width: number;
  height: number;
  imagePath: string;
}

export const LARGE_OBJECT_GRAPHICS_INFO: Record<LargeObjectGraphicsId, LargeObjectGraphicsInfo> = {
  'OBJ_EVENT_GFX_TRUCK': { width: 48, height: 48, imagePath: '/pokeemerald/graphics/object_events/pics/misc/truck.png' },
  'OBJ_EVENT_GFX_MR_BRINEYS_BOAT': { width: 32, height: 32, imagePath: '/pokeemerald/graphics/object_events/pics/misc/mr_brineys_boat.png' },
  'OBJ_EVENT_GFX_SS_TIDAL': { width: 96, height: 40, imagePath: '/pokeemerald/graphics/object_events/pics/misc/ss_tidal.png' },
  'OBJ_EVENT_GFX_SUBMARINE_SHADOW': { width: 88, height: 32, imagePath: '/pokeemerald/graphics/object_events/pics/misc/submarine_shadow.png' },
  'OBJ_EVENT_GFX_CABLE_CAR': { width: 64, height: 64, imagePath: '/pokeemerald/graphics/object_events/pics/misc/cable_car.png' },
};